requirements.txt
google-generativeai==0.3.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf==3.17.1
pydantic==2.4.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4